Feature vectors (X) and labels (y), GridSearchCV, selects the best model, evaluates its performance, and saves the trained model and class dictionary.
Feature vectors X and labels y are converted into numpy arrays for model training. The data is split into training and testing sets, and an SVM with an RBF kernel is trained and evaluated, To identify the optimal model, GridSearchCV is used to compare SVM, Random Forest, and Logistic Regression classifiers with various parameters. Confusion matrices and classification reports are generated to analyze performance. Finally, the best model and the class dictionary are saved for future use.